% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klustR.R, R/pacoplot.R, R/pcplot.R
\name{klustR-shiny}
\alias{klustR-shiny}
\alias{pacoplotOutput}
\alias{renderpacoplot}
\alias{pcplotOutput}
\alias{renderpcplot}
\title{Shiny bindings for klustR widgets}
\usage{
pacoplotOutput(outputId, width = "100\%", height = "400px")

renderpacoplot(expr, env = parent.frame(), quoted = FALSE)

pcplotOutput(outputId, width = "100\%", height = "400px")

renderpcplot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a networkD3 graph}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using klustR widgets within Shiny
applications and interactive Rmd documents.
}
