% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacoplot.R
\name{pacoplot}
\alias{pacoplot}
\title{Parallel Coordinates Plot for Clustering}
\usage{
pacoplot(data, clusters, colorScheme = "schemeCategory10", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A dataframe of numeric columns.}

\item{clusters}{A named integer matrix of clusters where names are the row names
of the above dataframe and integers are the integer value of the row's associated cluster.
This can be obtained from a function such as \code{ stats::kmeans()$cluster }.}

\item{colorScheme}{The color scheme of the PCA plot. May be a preconfigured D3 ordinal color scheme
or a vector of html colors (hex or named).}
}
\description{
Creates an interactive parallel coordinates plot detailing each
dimension and the cluster associated with each observation.
}
\details{
\itemize{
  \item Hover over lines to display row label
  \item Click on a line to fade out all lines except the associated cluster
  \item Click on another line to bold this line as well
  \item Clicking a second time on a line will fade it out
}

#' @examples
\dontrun{

df <- state.x77
clus <- kmeans(data_scaled, 5)$cluster
pacoplot(data = df, clusters = clus)

}
}
